+++ src/UIPlatform/Render/RenderData.h
@@ -15,6 +15,7 @@ namespace Meridian::Render
     {
         BrowserTransport browserTransport = BrowserTransport::SharedTexture;
         int cpuUploadFrameRate = 30;
+        bool gameDeviceNifRendering = false;
         ID3D11Device* device = nullptr;
         ID3D11DeviceContext3* deviceContext = nullptr;
         std::shared_ptr<::DirectX::CommonStates> commonStates = nullptr;